+++ First game/src/SpeedUp.java
@@ -17,11 +17,11 @@ public class SpeedUp extends GameObject{
 	}
 	public void render(Graphics g) {
 		g.setColor(Color.YELLOW);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 }
